test(stella-cli): witness the stale-lane diagnostic's own emission, and split it into a sibling module
#4153 asked for a decision plus, on the diagnostic-only path, a shipped
measurement: a stella-diag record at turn close when the board holds an
unowned in-progress task and the turn changed at least the threshold count
of files.

That record already exists (#4504's turn_files.rs::note_stale_lane), with
the exact predicate, level, and count-only fields the design pointer asks
for, registered in the diagnostic-codes gate and documented. The one gap
was that no test observed the record itself, only the pure decision
function behind it (stale_lane_open_tasks) - so a typo in a field key or
the record's code would have shipped green.

note_stale_lane now takes its Dx handle as a parameter instead of reaching
for the process-wide one, matching diag_boot's own stated convention for
code that is not spawned onto a drain. That is what lets a test pass a
capturing Dx and assert on the emitted record directly: task 1 in progress
plus three file writes emits agent.task.stale_lane with the right fields,
and a completed task emits nothing.

Moved into a new sibling module, turn_files/stale_lane.rs, because
turn_files.rs was already at 1473 of its 1500-line ceiling (AGENTS.md
Sec "God files") - the file's own module doc already explains why new
logic lands beside it rather than in it.

The design pointer named crates/stella-cli/src/agent/turn_close.rs as the
emit site. That function (turn_close::close_turn) settles the executions
row across every turn-closing driver, including the goal arcs; this
diagnostic is deliberately scoped narrower, to the two drivers that close
a turn through turn_files::close_turn_boundary, right beside the file-count
measurement (emit_shared_tree_changes) it needs. Moving it would either
duplicate that measurement or fold the goal arcs into a predicate whose own
doc comment explains why they are excluded. turn_files.rs already satisfies
the design's actual reasoning - a sibling of the near-ceiling agent.rs, not
agent.rs itself.

Decision (SCR-002, from the issue's design-pointer comment): the advisory
board plus #4152's refusal is the intended ceiling until a measured
frequency says otherwise. No injected reminder - #2412 measured board
traffic at 12-16% of a losing run's budget, and a reminder buys more of it.
Gap 2 (a lifecycle sentence in the system prompt) stays out for the same
reason. The sink-reachability gap that stops the diagnostic from producing
data today (Level::Debug, no sink captures it) is tracked separately as
#5075 and is out of scope here.

Witness: note_stale_lane_emits_only_for_the_walked_past_shape (new,
turn_files/stale_lane.rs) constructs a board with task 1 in progress and
calls note_stale_lane with 3 file writes, then asserts the captured Dx
holds a Debug-level agent.task.stale_lane record with files_changed=3,
open_tasks=2, threshold=3; a second call after completing the task asserts
no record. It could not compile against the old signature, let alone pass,
without this change - the old note_stale_lane took no Dx parameter at all.

Tests deleted: none. only_an_unowned_in_progress_card_over_the_file_threshold_is_recorded
moved verbatim (prose trimmed for the new file's empty prose-guard baseline)
from turn_files.rs into turn_files/stale_lane.rs.

Closes #4153
