+++ crates/stella-cli/src/turn_files.rs
@@ -77,6 +77,7 @@ use crate::config::Config;
 use crate::durability::{SessionDurability, UnmeasurableReason, WorktreeSnapshot};
 
 pub(crate) mod attribution;
+mod stale_lane;
 
 use attribution::Provenance;
 
@@ -307,7 +308,11 @@ pub(crate) fn close_turn_boundary(
     execution: Option<&(Arc<Store>, i64)>,
     outcome: &TurnOutcome,
 ) {
-    note_stale_lane(registry, emit_shared_tree_changes(cfg, tx, execution));
+    stale_lane::note_stale_lane(
+        &crate::diag_boot::dx(),
+        registry,
+        emit_shared_tree_changes(cfg, tx, execution),
+    );
     crate::agent::persistence::emit_run_complete_for_turn(tx, &cfg.model_id, outcome);
 }
 
@@ -339,92 +344,6 @@ pub(crate) fn close_turn_boundary_raw(
         execution,
         outcome,
     );
-}
-
-/// How many changed files make an untouched board worth a record.
-///
-/// One file is the ordinary shape of a turn that is genuinely still on the
-/// card it says it is on. Three is a guess, and it is meant to be: the point
-/// of the record is to find out what the real distribution looks like before
-/// anything is injected into the model's context.
-const STALE_LANE_FILES: usize = 3;
-
-/// Record a turn that ended with the lead's task lane still occupied while it
-/// changed files (#4153).
-///
-/// # What this does not claim
-///
-/// Not a defect signal. A turn that ends mid-card having edited three files is
-/// the ordinary shape of work in progress, and this fires on it. The thing
-/// worth knowing is the *rate*: #4152 closed the half of the divergence that
-/// has a correction point — `task_start` refuses a second unowned lane — and
-/// left the half that has none, where the agent never calls `task_start` at
-/// all and just starts editing. "These edits belong to a later step" is not
-/// decidable from the board, so the only honest first move is to measure how
-/// often a turn ends in that shape at all. `agent.turn_complete` is emitted
-/// once per turn, so the ratio is a join, not a second counter.
-///
-/// # Which drivers it covers
-///
-/// The two that close a turn through [`close_turn_boundary`]: `run_turn`
-/// (`stella run`, the plain REPL) and the deck's lead turn, which is the
-/// surface the divergence was reported on. The goal arcs and `run_resume`
-/// measure per round through [`emit_shared_tree_changes_raw`] and pay one
-/// terminator at the end (#4159), so a lane observation there is a question
-/// about a round rather than about a turn — a different reading, and one to
-/// take only if the numbers this produces say it is worth taking.
-///
-/// # Why counts only
-///
-/// A task id and a subject are model output, and `stella-diag`'s field types
-/// refuse both — deliberately, and the constraint is right here: a record
-/// naming the card the agent walked past would put model-authored text into
-/// the diagnostic plane for a signal that only needs an integer.
-fn note_stale_lane(registry: &stella_tools::ToolRegistry, files_changed: usize) {
-    let board = registry.task_board();
-    let board = board
-        .lock()
-        .unwrap_or_else(|poisoned| poisoned.into_inner());
-    let Some(open_tasks) = stale_lane_open_tasks(&board, files_changed) else {
-        return;
-    };
-    crate::diag_boot::dx().emit(Record::new(
-        Level::Debug,
-        "agent.task.stale_lane",
-        DIAG_TARGET,
-        Cx::EMPTY,
-        Fields::new()
-            .with("files_changed", files_changed as u64)
-            .with("open_tasks", open_tasks as u64)
-            .with("threshold", STALE_LANE_FILES as u64),
-    ));
-}
-
-/// The decision [`note_stale_lane`] makes, as a total function of the board
-/// and the count — `Some(open_tasks)` when the record is owed.
-///
-/// Split out because the emit half needs a `ToolRegistry`, a lock and a
-/// process-wide `Dx`, and none of those is the part that can be wrong. The
-/// part that can be wrong is *which* boards count, and it has two edges a
-/// reader would not guess: a delegated card is owned and does not occupy the
-/// lead's lane, and `open_tasks` counts pending cards too, because a plan the
-/// agent stopped walking is exactly the shape worth telling from a one-card
-/// board.
-fn stale_lane_open_tasks(
-    board: &stella_core::tasks::TaskBoard,
-    files_changed: usize,
-) -> Option<usize> {
-    if files_changed < STALE_LANE_FILES {
-        return None;
-    }
-    board.unowned_in_progress()?;
-    Some(
-        board
-            .items()
-            .iter()
-            .filter(|task| task.status.is_open())
-            .count(),
-    )
 }
 
 /// Measure the shared work tree once, then write both of this turn's file
@@ -697,62 +616,6 @@ mod tests {
                  surface while every other surface keeps working."
             );
         }
-    }
-
-    /// **Witness (#4153).** A turn that changed files while the lead's task
-    /// lane sat occupied is recorded, and the three boards that look like it
-    /// but are not are left alone.
-    ///
-    /// The delegated case is the one worth pinning: `task_assign` gives a card
-    /// an owner and deliberately does not route through `set_status`, so a
-    /// session that fanned work out to sub-agents has N in-progress cards and
-    /// is a perfectly honest board. A predicate that counted them would report
-    /// every delegating session and the measurement would be worthless.
-    #[test]
-    fn only_an_unowned_in_progress_card_over_the_file_threshold_is_recorded() {
-        use stella_core::tasks::TaskBoard;
-        use stella_protocol::event::TaskStatus;
-
-        let mut walked_past = TaskBoard::new();
-        walked_past.create("investigate the embedding pass", None, None);
-        walked_past.create("rewrite the chunker", None, None);
-        walked_past.set_status("1", TaskStatus::InProgress).unwrap();
-
-        assert_eq!(
-            stale_lane_open_tasks(&walked_past, STALE_LANE_FILES),
-            Some(2),
-            "an unowned card open across a file-changing turn is the shape #4153 is about"
-        );
-        assert_eq!(
-            stale_lane_open_tasks(&walked_past, STALE_LANE_FILES - 1),
-            None,
-            "under the threshold a turn is ordinary work in progress, not a signal"
-        );
-
-        let mut delegated = TaskBoard::new();
-        delegated.create("run the fan-out", None, None);
-        delegated.assign("1", "sub:worker-a").unwrap();
-        assert_eq!(
-            stale_lane_open_tasks(&delegated, STALE_LANE_FILES * 10),
-            None,
-            "a delegated card carries an owner and does not occupy the lead's lane"
-        );
-
-        let mut finished = TaskBoard::new();
-        finished.create("rewrite the chunker", None, None);
-        finished.set_status("1", TaskStatus::InProgress).unwrap();
-        finished.set_status("1", TaskStatus::Completed).unwrap();
-        assert_eq!(
-            stale_lane_open_tasks(&finished, STALE_LANE_FILES * 10),
-            None,
-            "a board the agent kept current is the whole point and must stay silent"
-        );
-
-        assert_eq!(
-            stale_lane_open_tasks(&TaskBoard::new(), STALE_LANE_FILES * 10),
-            None,
-            "a session with no board at all cannot have walked past a card"
-        );
     }
 
     /// **Witness (#4175).** Every driver that opens a turn's event stream
