+++ docs/reference/diagnostics.md
@@ -413,7 +413,7 @@ A subagent event crossed the stream. Placement only; the subagent's own run has 
 
 <!-- facts:begin — generated by `make diag-reference`; do not edit by hand -->
 - **Level:** `debug`
-- **Emitted from:** `crates/stella-cli/src/turn_files.rs`
+- **Emitted from:** `crates/stella-cli/src/turn_files/stale_lane.rs`
 - **Fields:** `files_changed`, `open_tasks`, `threshold`
 <!-- facts:end -->
